fix: Make leader election select from the whole Committee (MystenLabs/narwhal#762)
* fix: Make leader election select from the whole Committee

Fixes MystenLabs/narwhal#755

* fix: Make tests of binomial result more strict

* fix: make the leader see a strict u64 (from usize)

* fix: consensus behavior under tests

+++ narwhal/primary/src/proposer.rs
@@ -137,7 +137,7 @@ impl Proposer {
     // Main loop listening to incoming messages.
     /// Update the last leader certificate. This is only relevant in partial synchrony.
     fn update_leader(&mut self) -> bool {
-        let leader_name = self.committee.leader(self.round as usize);
+        let leader_name = self.committee.leader(self.round);
         self.last_leader = self
             .last_parents
             .iter()
